* V5-8-patches:
  Cygwin, MIB-II: Fix a few compiler warnings
  Cygwin, IF-MIB: Temporarily disable the iphlpapi implementation
  Windows, IF-MIB: Minimize #include directives
  Net-SNMP include files: Minimize #include directives
  include/net-snmp/system/cygwin.h: Simplify this file
  ci, .travis.yml: Use OS/X 10.14
  Net-SNMP headers: Move in_addr_t definition into <net-snmp/net-snmp-config.h>
  Net-SNMP headers: Only include <sys/time.h> if necessary
  IF-MIB: Unbreak the MinGW32 build
  apps/Makefile.in: Make this Makefile again compatible with non-GNU make implementations

+++ agent/mibgroup/mibII/data_access/at_iphlpapi.c
@@ -385,7 +385,7 @@ write_arp(int action,
             if (!create_flag) {
                 if ((status = SetIpNetEntry(oldarp_row)) != NO_ERROR) {
                     snmp_log(LOG_ERR, "Error in case UNDO, status %u\n",
-                             status);
+                             (unsigned int)status);
                     retval = SNMP_ERR_UNDOFAILED;
                 }
             }
@@ -396,7 +396,7 @@ write_arp(int action,
                 if ((status = SetIpNetEntry(arp_row)) != NO_ERROR) {
                     snmp_log(LOG_ERR,
                              "Error while deleting added row, status %u\n",
-                             status);
+                             (unsigned int)status);
                     retval = SNMP_ERR_UNDOFAILED;
                 }
             }
@@ -416,7 +416,7 @@ write_arp(int action,
                 if ((status = CreateIpNetEntry(arp_row)) != NO_ERROR) {
                     snmp_log(LOG_ERR,
                              "Inside COMMIT: CreateIpNetEntry failed, status %u\n",
-                             status);
+                             (unsigned int)status);
                     retval = SNMP_ERR_COMMITFAILED;
                 }
             } else {
